+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - selenium 3.3.3 -> 3.7.0
 - requests 2.13.0 -> >=2.18.2
 - ndg-httpsclient 0.4.2 -> 0.4.3
+- [ChromeDriver](https://sites.google.com/a/chromium.org/chromedriver) 2.29 -> 2.33
 
 ### Removed